# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation for SMIL",
        "contact": {
            "email": "muhammad.raflysadewa.tik17@mhsw.pnj.ac.id"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/lokasi": {
            "get": {
                "tags": [
                    "Lokasi Penyimpanan"
                ],
                "summary": "Return List of Lokasi",
                "operationId": "getAllLokasi",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lokasi Penyimpanan"
                ],
                "summary": "Create New Lokasi Penyimpanan",
                "operationId": "createNewLokasiPenyimpanan",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreLokasiPenyimpananRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LokasiPenyimpananDetailResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/filter/lokasi": {
            "post": {
                "tags": [
                    "Lokasi Penyimpanan"
                ],
                "summary": "Return List of Lokasi Penyimpanan by Filter",
                "operationId": "getAllLokasiPenyimpananByFilter",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FilterLokasiPenyimpananRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LokasiPenyimpananListResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/lokasi/available/{totalNeed}": {
            "get": {
                "tags": [
                    "Lokasi Penyimpanan"
                ],
                "summary": "Get List Available Lokasi",
                "operationId": "getAllAvailabelLokasi",
                "parameters": [
                    {
                        "name": "totalNeed",
                        "in": "path",
                        "description": "Jumlah Kapasitas yang dibutuhkan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LokasiPenyimpananDetailResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/lokasi/{id}": {
            "get": {
                "tags": [
                    "Lokasi Penyimpanan"
                ],
                "summary": "Get Lokasi Detail Information",
                "operationId": "getAllLokasiById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lokasi ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LokasiPenyimpananDetailResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Lokasi Penyimpanan"
                ],
                "summary": "Update Existed Lokasi Penyimpanan",
                "operationId": "updateExistedLokasiPenyimpanan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lokasi Penyimpanan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreLokasiPenyimpananRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LokasiPenyimpananDetailResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lokasi Penyimpanan"
                ],
                "summary": "Delete Existed Lokasi Penyimpanan",
                "operationId": "deleteExistingLokasi Penyimpanan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lokasi Penyimpanan ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LokasiPenyimpanan": {
                "title": "LokasiPenyimpanan",
                "description": "Lokasi Penyimpanan Model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "lokasi_name": {
                        "title": "Lokasi Penyimpanan Name",
                        "description": "Name of the new Lokasi Penyimpanan",
                        "type": "string",
                        "example": "Lemari A"
                    },
                    "total_capacity": {
                        "title": "Total Capacity",
                        "description": "Total Capacity of the new Lokasi Penyimpanan",
                        "type": "integer",
                        "example": 100
                    },
                    "available_capacity": {
                        "title": "Available Capacity",
                        "description": "Available Capacity of the new Lokasi Penyimpanan",
                        "type": "integer",
                        "example": 100
                    },
                    "stored_capacity": {
                        "title": "Store Capacity",
                        "description": "Store Capacity of the new Lokasi Penyimpanan",
                        "type": "integer",
                        "example": 100
                    },
                    "path_qrcode": {
                        "title": "Path QR Code",
                        "description": "Path QR Code of the new Lokasi Penyimpanan",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created Date",
                        "description": "Date of data been created",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "deleted_at": {
                        "title": "Deleted at",
                        "description": "Deleted at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Lokasi Penyimpanan"
                }
            },
...
            "Response": {
                "title": "Response",
                "description": "Response Model",
                "properties": {
                    "code": {
                        "title": "Code",
                        "description": "Code Response",
                        "type": "integer",
                        "format": "int64",
                        "example": 200
                    },
                    "status": {
                        "title": "Status Response",
                        "description": "Status Response",
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "title": "Message Response",
                        "description": "Message Response",
                        "type": "string",
                        "example": "Response Berhasil"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Response"
                }
            },
...
            "FilterLokasiPenyimpananRequest": {
                "title": "Filter Lokasi Penyimpanan request",
                "description": "Filter Lokasi Penyimpanan request body data",
                "properties": {
                    "page_size": {
                        "title": "page_size",
                        "description": "Page Size of Filter",
                        "type": "integer",
                        "example": 5
                    },
                    "sort_by": {
                        "title": "sort_by",
                        "description": "Sort Filter By Attributes",
                        "type": "string",
                        "example": "id"
                    },
                    "sort_direction": {
                        "title": "sort_direction",
                        "description": "Sort Direction of Filter",
                        "type": "string",
                        "example": "ASC"
                    },
                    "lokasi_name": {
                        "title": "lokasi_name",
                        "description": "Filter Nama Lokasi",
                        "type": "string"
                    },
                    "total_capacity": {
                        "title": "total_capacity",
                        "description": "Filter Kapasitas Total",
                        "type": "integer"
                    },
                    "available_capacity": {
                        "title": "available_capacity",
                        "description": "Filter Kapasitas Tersedia",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "StoreLokasiPenyimpananRequest": {
                "title": "Store Lokasi Penyimpanan Request",
                "description": "Store Lokasi Penyimpanan Request body data",
                "required": [
                    "lokasi_name",
                    "total_capacity"
                ],
                "properties": {
                    "lokasi_name": {
                        "title": "lokasi_name",
                        "description": "Lokasi Name",
                        "type": "string",
                        "example": "Lemari A"
                    },
                    "total_capacity": {
                        "title": "total_capacity",
                        "description": "Total Capacity",
                        "type": "integer",
                        "example": 100
                    },
                    "available_capacity": {
                        "title": "available_capacity",
                        "description": "Available Capacity",
                        "type": "integer",
                        "example": 100
                    },
                    "stored_capacity": {
                        "title": "stored_capacity",
                        "description": "Store Capacity",
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
...
            "LokasiPenyimpananDetailResource": {
                "title": "LokasiPenyimpananDetailResource",
                "description": "Lokasi Penyimpanan Resource",
                "properties": {
                    "response": {
                        "$ref": "#/components/schemas/Response"
                    },
                    "data": {
                        "title": "Data",
                        "description": "Data Lokasi Penyimpanan Wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LokasiPenyimpanan"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LokasiPenyimpananDetailResource"
                }
            },
            "LokasiPenyimpananListResource": {
                "title": "LokasiPenyimpananListResource",
                "description": "Lokasi Penyimpanan Resource",
                "properties": {
                    "result": {
                        "title": "Result",
                        "description": "Result Lokasi Penyimpanan Wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LokasiPenyimpanan"
                        }
                    },
                    "page": {
                        "$ref": "#/components/schemas/Page"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LokasiPenyimpananListResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}